requirements.txt
getkey
eclipse-zenoh
RPi.GPIO
pygame